Rename dor-workflow-service to dor-workflow-client
This matches the pattern established by dor-services-client
