+++ .rubocop_todo.yml
@@ -1,20 +1,10 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-04-03 18:57:07 -0500 using RuboCop version 0.63.1.
+# on 2019-04-04 09:19:56 -0500 using RuboCop version 0.63.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
-
-# Offense count: 1
-Lint/AmbiguousRegexpLiteral:
-  Exclude:
-    - 'spec/workflow_service_spec.rb'
-
-# Offense count: 1
-Lint/UselessAssignment:
-  Exclude:
-    - 'lib/dor/services/workflow_service.rb'
 
 # Offense count: 1
 Metrics/AbcSize:
@@ -34,10 +24,3 @@ Metrics/MethodLength:
 # Configuration parameters: CountKeywordArgs.
 Metrics/ParameterLists:
   Max: 6
-
-# Offense count: 1
-# Configuration parameters: ExpectMatchingDefinition, Regex, IgnoreExecutableScripts, AllowedAcronyms.
-# AllowedAcronyms: CLI, DSL, ACL, API, ASCII, CPU, CSS, DNS, EOF, GUID, HTML, HTTP, HTTPS, ID, IP, JSON, LHS, QPS, RAM, RHS, RPC, SLA, SMTP, SQL, SSH, TCP, TLS, TTL, UDP, UI, UID, UUID, URI, URL, UTF8, VM, XML, XMPP, XSRF, XSS
-Naming/FileName:
-  Exclude:
-    - 'lib/dor-workflow-service.rb'
