--- lib/dor-workflow-service.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require 'dor/workflow_version'
-require 'dor/services/workflow_service'
